Changed int to size_t for comparison

+++ ft_strdup.c
@@ -6,7 +6,7 @@
 /*   By: pnguyen- <pnguyen-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/06 18:33:38 by pnguyen-          #+#    #+#             */
-/*   Updated: 2023/11/11 16:58:12 by pnguyen-         ###   ########.fr       */
+/*   Updated: 2023/11/11 17:07:16 by pnguyen-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 char	*ft_strdup(const char s[])
 {
 	char	*str;
-	int		i;
+	size_t	i;
 	size_t	len;
 
 	len = ft_strlen(s);
